% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{A simple function for loading packages within a package development environment}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{a charater/list of packages that all have to be in CRAN}
}
\description{
This function allows a user a load a package while guaranteeing code will install missing packages
}
