% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterextractionfunctions.R
\name{parallel_extract}
\alias{parallel_extract}
\title{Function to extract rasters}
\usage{
parallel_extract(shp_dt, raster_list, fun_list, numCores)
}
\description{
Function to extract rasters
}
